Add speaker extraction and chunked multithreading
