+++ 0-prompts/determine-and-summarize-speakers-prompt.md
@@ -1,0 +1,19 @@
+You will receive a raw story that will be turned into an audiobook.
+Identify every speaker or entity that has spoken text in the story. Include narrator voices, named characters and any other entity that should speak with a distinct voice.
+For each of them provide the following information strictly in this format:
+
+# Speaker {number}
+Name: <short name or identifier>
+Description: <short description of the character>
+Personality & behaviour: <key traits>
+Voice type: <choose from the list below>
+
+Use each speaker only once and do not omit anyone. Reply only with the list of speakers.
+
+Available voices:
+{voices}
+
+Story text:
+```
+{text}
+```
