+++ audiobook-tts/src/split_voice_lines.py
@@ -1,5 +1,6 @@
 import os
 import re
+from text_utils import TextUtils
 import json
 import yaml
 import argparse
@@ -15,8 +16,6 @@ zonos_min_chars = config["models"]["zonos"]["min_characters"]
 zonos_max_chars = config["models"]["zonos"]["max_characters"]
 
 # --- Splitting Functions ---
-
-ABBREVS = {"dr", "mr", "mrs", "ms", "etc"}  # Lowercase set for quick check
 
 def parse_speakers(full_text, default_voice):
     """
@@ -58,132 +57,6 @@ def parse_speakers(full_text, default_voice):
             segments.append((default_voice, final_text, ""))
     return segments
 
-def _split_regular_text(text):
-    sentences = []
-    buffer = []
-    i = 0
-    n = len(text)
-    while i < n:
-        ch = text[i]
-        buffer.append(ch)
-        if ch in {'.', '!', '?'}:
-            next_char = text[i + 1] if (i + 1 < n) else ''
-            if next_char.isspace() or i == n - 1:
-                temp_str = "".join(buffer)
-                match = re.search(r"\b([A-Za-z]+)\.$", temp_str)
-                if match:
-                    word_before_dot = match.group(1).lower()
-                    if word_before_dot in ABBREVS:
-                        pass
-                    else:
-                        sentences.append(temp_str.strip())
-                        buffer = []
-                        while i + 1 < n and text[i + 1].isspace():
-                            i += 1
-                else:
-                    sentences.append(temp_str.strip())
-                    buffer = []
-                    while i + 1 < n and text[i + 1].isspace():
-                        i += 1
-        i += 1
-    leftover = "".join(buffer).strip()
-    if leftover:
-        sentences.append(leftover)
-    return sentences
-
-def initial_split_into_sentences(text):
-    text = re.sub(r'\s*</pause>', '', text)
-    parts = re.split(r'(<pause\s+duration="(?:short|long)"\s*/?>)', text)
-    sentences = []
-    for i, part in enumerate(parts):
-        if i % 2 == 0:
-            sub_sentences = _split_regular_text(part)
-            if sub_sentences:
-                sentences.extend(sub_sentences)
-        else:
-            m = re.search(r'duration="(short|long)"', part)
-            duration = m.group(1) if m else "short"
-            if sentences:
-                sentences[-1] = sentences[-1].strip() + " ||PAUSE:" + duration + "||"
-            else:
-                sentences.append("||PAUSE:" + duration + "||")
-    return sentences
-
-def split_on_delims(long_text, max_len):
-    def splitter(text, delimiters):
-        pieces = [text]
-        for delim in delimiters:
-            new_pieces = []
-            for p in pieces:
-                if len(p) <= max_len:
-                    new_pieces.append(p)
-                else:
-                    sub_parts = p.split(delim)
-                    for sp in sub_parts:
-                        sp = sp.strip()
-                        if sp:
-                            new_pieces.append(sp)
-            pieces = new_pieces
-        return pieces
-
-    candidate_delims = [" - ", ", "]
-    splitted = splitter(long_text, candidate_delims)
-    final_list = []
-    for piece in splitted:
-        if len(piece) > max_len:
-            final_list.extend(force_split_long_text(piece, max_chars=max_len))
-        else:
-            final_list.append(piece)
-    return final_list
-
-def force_split_long_text(text, max_chars):
-    if len(text) <= max_chars:
-        return [text]
-
-    splits = []
-    start = 0
-    length = len(text)
-    while start < length:
-        end = start + max_chars
-        if end >= length:
-            splits.append(text[start:].strip())
-            break
-        space_index = text.rfind(" ", start, end)
-        if space_index == -1 or space_index < start:
-            splits.append(text[start:end].strip())
-            start = end
-        else:
-            splits.append(text[start:space_index].strip())
-            start = space_index + 1
-    return splits
-
-def combine_into_blocks(sentences, min_chars, max_chars):
-    blocks = []
-    current = ""
-    for sentence in sentences:
-        m = re.search(r'\|\|PAUSE:(short|long)\|\|$', sentence)
-        if m:
-            pause = m.group(1)
-            sentence_text = sentence[:sentence.rfind("||PAUSE:")].strip()
-            if current:
-                blocks.append((current.strip(), ""))
-                current = ""
-            blocks.append((sentence_text, pause))
-        else:
-            if not current:
-                current = sentence
-            else:
-                if len(current) + 1 + len(sentence) <= max_chars:
-                    current += " " + sentence
-                else:
-                    if len(current) >= min_chars:
-                        blocks.append((current.strip(), ""))
-                        current = sentence
-                    else:
-                        current += " " + sentence
-    if current.strip():
-        blocks.append((current.strip(), ""))
-    return blocks
 
 # --- Main function: split_voice_lines ---
 
@@ -233,15 +106,15 @@ def split_voice_lines(input_file=None, input_dir="2-annotated-text", temp_folder
             model_min_chars = zonos_min_chars if model == "zonos" else kokoro_min_chars
             model_max_chars = zonos_max_chars if model == "zonos" else kokoro_max_chars
 
-            sentences = initial_split_into_sentences(segment_text)
+            sentences = TextUtils.initial_split_into_sentences(segment_text)
             refined = []
             for s in sentences:
                 if len(s) > model_max_chars:
-                    refined.extend(split_on_delims(s, max_len=model_max_chars))
+                    refined.extend(TextUtils.split_on_delims(s, max_len=model_max_chars))
                 else:
                     refined.append(s)
 
-            blocks = combine_into_blocks(refined, min_chars=model_min_chars, max_chars=model_max_chars)
+            blocks = TextUtils.combine_into_blocks(refined, min_chars=model_min_chars, max_chars=model_max_chars)
             for block_text, pause in blocks:
                 block_text = block_text.strip()
                 if block_text:
